requirements.txt
PyYAML==3.12
mysql-connector-python-rf==2.2.2

